+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
+- Set widget like as Text, Icon
+- Implement onTap()
+- Implement onLongPress()
+- Implement onLongPressCancel()
+- Input delay duration as seconds
+- Add decoration property
 
-* TODO: Describe initial release.